# specs/002-admin-dashboard/contracts/api.yaml
openapi: 3.0.3
info:
  title: Admin Dashboard API
  description: API REST pour l'application d'administration de la plateforme dropshipping
  version: 1.0.0
servers:
  - url: http://localhost:8000/admin
    description: Serveur de développement
  - url: https://api.dropshipping.local/admin
    description: Serveur de production
...
paths:
  /auth/login:
    post:
      tags:
        - auth
      summary: Connexion administrateur
      description: Authentifie un administrateur et retourne un token JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: admin@dropshipping.local
                password:
                  type: string
                  format: password
                  example: admin123
      responses:
        '200':
          description: Connexion réussie
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: Token JWT pour authentification
                  expires_at:
                    type: string
                    format: date-time
                    description: Date d'expiration du token
                  admin:
                    $ref: '#/components/schemas/AdminUser'
        '401':
          description: Identifiants invalides
        '403':
          description: Compte désactivé
  /auth/logout:
    post:
      tags:
        - auth
      summary: Déconnexion administrateur
      description: Révoque la session actuelle
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Déconnexion réussie
        '401':
          description: Non authentifié
  /auth/me:
    get:
      tags:
        - auth
      summary: Informations administrateur actuel
      description: Retourne les informations de l'administrateur authentifié
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Informations administrateur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUser'
        '401':
          description: Non authentifié
...
  /campaigns:
    get:
      tags:
        - campaigns
      summary: Liste des campagnes
      description: Retourne la liste des campagnes avec filtres
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [paused, active, completed]
          description: Filtrer par statut
        - name: platform
          in: query
          schema:
            type: string
            enum: [meta, tiktok]
          description: Filtrer par plateforme
      responses:
        '200':
          description: Liste des campagnes
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Campaign'
                  total:
                    type: integer
        '401':
          description: Non authentifié
  /campaigns/{campaign_id}/activate:
    post:
      tags:
        - campaigns
      summary: Activer une campagne
      description: Active une campagne en pause
      security:
        - bearerAuth: []
      parameters:
        - name: campaign_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Campagne activée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Campaign'
        '400':
          description: Campagne déjà active ou erreur
        '401':
          description: Non authentifié
        '404':
          description: Campagne non trouvée
  /campaigns/{campaign_id}/deactivate:
    post:
      tags:
        - campaigns
      summary: Désactiver une campagne
      description: Met en pause une campagne active
      security:
        - bearerAuth: []
      parameters:
        - name: campaign_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Campagne désactivée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Campaign'
        '400':
          description: Campagne déjà en pause ou erreur
        '401':
          description: Non authentifié
        '404':
          description: Campagne non trouvée
...
  /actions/scoring/calculate:
    post:
      tags:
        - actions
      summary: Calculer les scores
      description: Déclenche manuellement le calcul des scores pour les produits non scorés
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Calcul des scores démarré
          content:
            application/json:
              schema:
                type: object
                properties:
                  job_id:
                    type: string
                    description: Identifiant du job de scoring
                  status:
                    type: string
                    enum: [queued, started]
        '401':
          description: Non authentifié
  /actions/creatives/generate/{product_id}:
    post:
      tags:
        - actions
      summary: Générer les créatives pour un produit
      description: Déclenche manuellement la génération de créatives pour un produit
      security:
        - bearerAuth: []
      parameters:
        - name: product_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Génération de créatives démarrée
          content:
            application/json:
              schema:
                type: object
                properties:
                  job_id:
                    type: string
                    description: Identifiant du job de génération
                  status:
                    type: string
                    enum: [queued, started]
        '401':
          description: Non authentifié
        '404':
          description: Produit non trouvé
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    AdminUser:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
          format: email
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        last_login_at:
          type: string
          format: date-time
          nullable: true
...
    Campaign:
      type: object
      properties:
        id:
          type: integer
        product_id:
          type: integer
        platform:
          type: string
          enum: [meta, tiktok]
        campaign_name:
          type: string
        status:
          type: string
          enum: [paused, active, completed]
        daily_budget:
          type: integer
          description: Budget quotidien en centimes
        created_at:
          type: string
          format: date-time
        activated_at:
          type: string
          format: date-time
          nullable: true
...